[feat] 게시글 view count 쿼리 수정

+++ kitten-api/src/main/java/kitten/diy/api/application/port/in/query/data/BoardInfoData.java
@@ -41,7 +41,7 @@ public record BoardInfoData(
                 .type(data.type())
                 .createTime(data.createTime())
                 .likeCount(data.likeCount().intValue())
-                .viewCount(data.viewCount())
+                .viewCount(data.viewCount().intValue())
                 .tags(tags)
                 .build();
     }
